Rename subdir and abstract shared components
